refactor(cpuhealth): split capacity.go by the source each reader reads
capacity.go carried the sampler vocabulary and a reader that pulled from
four separate sources in one file: /proc/stat for the host busy and steal
jiffies, the cgroup's effective cpuset, cpu.pressure for PSI, and the
virtualization evidence spread across /proc/cpuinfo and two DMI files.
Each source now has its own file, and Scope, Sample, Sampler and
DeriveEnvironment move to sample.go so the vocabulary is readable without
scrolling past the reading code.

capacity.go keeps the cgroup sampler itself: NewCgroupSampler, the
cgroupSampler struct, Read, and the cpu.stat reading that Read depends on
directly (readStat and parseCounter).

This is a pure move. Every declaration travels with its doc comment; no
renames, no comment edits, no behaviour change. Verified: go doc -all -u
over the package hashes identically before and after, the suite still runs
80 of 80 specs, and every moved line is byte-identical to its original.
